BackEnd/requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.7
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2==2.9.5
psycopg2-binary
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0

